.......... [ZBXNEXT-1096] fixed warnings, include in db patch

+++ src/zabbix_server/poller/async_agent.c
@@ -314,7 +314,7 @@ int	zbx_async_check_agent(zbx_dc_item_t *item, AGENT_RESULT *result,  zbx_async_
 		agent_context->server_name = NULL;
 #endif
 	zbx_socket_clean(&agent_context->s);
-	zbx_tcp_send_context_init(agent_context->item.key, strlen(agent_context->item.key), timeout_sec,
+	zbx_tcp_send_context_init(agent_context->item.key, strlen(agent_context->item.key), (size_t)timeout_sec,
 		ZBX_TCP_PROTOCOL, &agent_context->tcp_send_context);
 
 	agent_context->step = ZABBIX_AGENT_STEP_CONNECT_WAIT;
